infra/blob-seed/requirements.txt
numpy==1.26.4
rasterio==1.3.10
rio-cogeo==5.3.3
azure-storage-blob==12.22.0
